requirements.txt
django
python-social-auth
django-bower
requests
djangorestframework
django-extensions
Pillow
django-extra-fields
pyopenssl
